src/complex.rs: moved code into `src/plugin_policy/`
