+++ src/lib.rs
@@ -1,5 +1,5 @@
-mod complex;
 pub mod policy;
+pub mod plugin_policy;
 
 pub use policy::*;
